refactor: split proposal page into smaller components
